refactor: break up test specs further
