+++ packages/date-picker/src/DatePicker/DatePickerMenu/DatePickerMenu.spec.tsx
@@ -7,7 +7,6 @@ import {
   Month,
   newUTC,
 } from '@leafygreen-ui/date-utils';
-import { mockTimeZone, testTimeZones } from '@leafygreen-ui/date-utils/testing';
 
 import {
   SharedDatePickerProvider,
@@ -216,46 +215,5 @@ describe('packages/date-picker/date-picker-menu', () => {
         });
       });
     });
-
-    describe.each(testTimeZones)(
-      'when system time is in $tz',
-      ({ tz, UTCOffset }) => {
-        describe.each([
-          { tz: undefined, UTCOffset: undefined },
-          ...testTimeZones,
-        ])('and timeZone prop is $tz', prop => {
-          const elevenLocal = 23 - (prop.UTCOffset ?? UTCOffset);
-          const midnightLocal = 0 - (prop.UTCOffset ?? UTCOffset);
-          const dec24Local = newUTC(2023, Month.December, 24, elevenLocal, 59);
-          const dec25Local = newUTC(2023, Month.December, 25, midnightLocal, 0);
-          const dec24ISO = '2023-12-24';
-          const dec25ISO = '2023-12-25';
-          const ctx = {
-            timeZone: prop?.tz,
-          };
-
-          beforeEach(() => {
-            jest.setSystemTime(dec24Local);
-            mockTimeZone(tz, UTCOffset);
-          });
-          afterEach(() => {
-            jest.restoreAllMocks();
-          });
-
-          test('when date changes, cell marked as `current` updates', () => {
-            const { getCellWithISOString, rerenderDatePickerMenu } =
-              renderDatePickerMenu(null, null, ctx);
-            const dec24Cell = getCellWithISOString(dec24ISO);
-            expect(dec24Cell).toHaveAttribute('aria-current', 'true');
-
-            jest.setSystemTime(dec25Local);
-
-            rerenderDatePickerMenu();
-            const dec25LocalCell = getCellWithISOString(dec25ISO);
-            expect(dec25LocalCell).toHaveAttribute('aria-current', 'true');
-          });
-        });
-      },
-    );
   });
 });
